+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy/A32NX_FMCMainDisplay.ts
@@ -1,5 +1,5 @@
 // @ts-strict-ignore
-// Copyright (c) 2021-2023, 2025 FlyByWire Simulations
+// Copyright (c) 2021-2023, 2025-2026 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
 import {
@@ -58,7 +58,7 @@ import { A320_Neo_CDU_MainDisplay } from './A320_Neo_CDU_MainDisplay';
 import { FmsDisplayInterface } from '@fmgc/flightplanning/interface/FmsDisplayInterface';
 import { FmsError, FmsErrorType } from '@fmgc/FmsError';
 import { FmsDataInterface } from '@fmgc/flightplanning/interface/FmsDataInterface';
-import { BitFlags, EventBus, Subscription, SimVarValueType } from '@microsoft/msfs-sdk';
+import { BitFlags, EventBus, Subscription, SimVarValueType, ClockEvents, ConsumerSubject } from '@microsoft/msfs-sdk';
 import { AdfRadioTuningStatus, MmrRadioTuningStatus, VorRadioTuningStatus } from '@fmgc/navigation/NavaidTuner';
 import { Coordinates } from '@fmgc/flightplanning/data/geo';
 import { FmsFormatters } from './FmsFormatters';
@@ -128,7 +128,6 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
   public unconfirmedV2Speed = undefined;
   public _toFlexChecked = true;
   private toRunway = undefined;
-  private _debug = undefined;
   public isDestEfobAmber = false;
   private isBelowMinDestFobForTwoMinutes?: NXLogicConfirmNode;
   private shouldShowBelowMinDestEfobMessage = false;
@@ -322,9 +321,18 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     SimVarValueType.Enum,
   );
 
+  private readonly sub = this.bus.getSubscriber<ClockEvents>();
+
+  /** Simulation time in milliseconds since the UNIX epoch (JS timestamp). Hint: this clock is affected by sim rate. */
+  private readonly simTime = ConsumerSubject.create(this.sub.on('simTime'), 0);
+
+  /** Paused if ETT is expired or does not exist */
+  private readonly ettCheckSub = this.sub.on('simTime').onlyAfter(1000).handle(this.checkEttExpired.bind(this), true);
+
+  private static readonly MILISECONDS_IN_DAY = 86400000;
+
   constructor(public readonly bus: EventBus) {
     FMCMainDisplay.DEBUG_INSTANCE = this;
-
     this.currFlightPlanService.createFlightPlans();
     this.currNavigationDatabaseService.activeDatabase = this.navigationDatabase;
   }
@@ -457,7 +465,6 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     this.unconfirmedV2Speed = undefined;
     this._toFlexChecked = true;
     this.toRunway = '';
-    this._debug = 0;
     this.isDestEfobAmber = false;
     this.isBelowMinDestFobForTwoMinutes = new NXLogicConfirmNode(120_000, true);
     this.shouldShowBelowMinDestEfobMessage = false;
@@ -603,6 +610,8 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     this.toSpeedsChecks();
 
     this.setRequest('FMGC');
+
+    this.clearEtt();
   }
 
   public onUpdate(deltaTime: number) {
@@ -617,10 +626,6 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     updateComponents(deltaTime);
 
     this.isTrueRefMode = SimVar.GetSimVarValue('L:A32NX_FMGC_TRUE_REF', 'boolean');
-
-    if (this._debug++ > 180) {
-      this._debug = 0;
-    }
     const flightPhaseManagerDelta = this.flightPhaseUpdateThrottler.canUpdate(deltaTime);
     if (flightPhaseManagerDelta !== -1) {
       this.flightPhaseManager.shouldActivateNextPhase(flightPhaseManagerDelta);
@@ -715,6 +720,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
 
     switch (nextPhase) {
       case FmgcFlightPhase.Takeoff: {
+        this.clearEtt();
         this._destDataChecked = false;
         if (plan.performanceData.accelerationAltitude.get() === null) {
           // it's important to set this immediately as we don't want to immediately sequence to the climb phase
@@ -2221,6 +2227,7 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     }
 
     this.atsu.resetAtisAutoUpdate();
+    this.clearEtt(forPlan);
 
     await this.flightPlanService.newCityPair(from, to, undefined, forPlan);
   }
@@ -5444,7 +5451,16 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
     const phase = this.getFlightPhase();
 
     if (phase === FmgcFlightPhase.Preflight || phase === FmgcFlightPhase.Done) {
+      const ett = this.flightPlanService.active.performanceData.estimatedTakeoffTime.get();
+      if (ett !== null && this.ettCheckSub.isPaused) {
+        if (!this.checkEttExpired()) {
+          this.ettCheckSub.resume();
+        }
+      }
       this.addMessageToQueue(NXSystemMessages.checkToData);
+    } else {
+      // Clear ETT if we are beyond the preflight phase
+      this.clearEtt();
     }
 
     if (zfwDiff !== null && zfwDiff > 5) {
@@ -5628,6 +5644,148 @@ export abstract class FMCMainDisplay implements FmsDataInterface, FmsDisplayInte
 
   getPerformanceFactorPercent(): number | null {
     return null; // TODO implement with PERF FACTOR in AC STATUS page
+  }
+
+  /**
+   * Sets the estimated takeoff time on the specified flightplan.
+   * @param text estimated takeoff time text. Must be in HHMMSS format.
+   * @param forPlan flightplan to set ETT on
+   * @returns void
+   */
+  setEstimatedTakeoffTime(text: string, forPlan: number): void {
+    const flightplan = this.flightPlanService.get(forPlan);
+    const activeOrCopyOfActive = flightplan.isActiveOrCopiedFromActive();
+
+    // If its not active or copy of active allow setting in any phase
+    if (!activeOrCopyOfActive || this.getFlightPhase() === FmgcFlightPhase.Preflight) {
+      if (text == Keypad.clrValue) {
+        // If ETT has expired we don't show it on the RTA page, as such, ignore clearing if thats the case
+        if (
+          flightplan.performanceData.estimatedTakeoffTime.get() !== null &&
+          !flightplan.performanceData.estimatedTakeoffTimeExpired.get()
+        ) {
+          this.clearEtt(forPlan);
+        } else {
+          this.setScratchpadMessage(NXSystemMessages.notAllowed);
+        }
+      } else {
+        // Accept HHMM or HHMMSS
+        const matches = /^(\d{1,2})(\d{2})(\d{0,2})$/.exec(text);
+
+        if (matches) {
+          const hours = parseInt(matches[1]);
+          const minutes = parseInt(matches[2]);
+          const seconds = matches[3] ? parseInt(matches[3]) : 0;
+          if (minutes > 59 || hours > 23 || seconds > 59) {
+            this.setScratchpadMessage(NXSystemMessages.entryOutOfRange);
+          } else {
+            const currentTimeStamp = this.simTime.get();
+            const zuluTimeMilliseconds = currentTimeStamp % FMCMainDisplay.MILISECONDS_IN_DAY;
+            const ettInMilliseconds =
+              (FmsFormatters.hoursToSeconds(hours) + FmsFormatters.minutesToSeconds(minutes) + seconds) * 1000;
+            let ettMillisecondsFromNow = ettInMilliseconds;
+
+            // If ETT is less than current time, assume it is for the next day
+            if (ettInMilliseconds < zuluTimeMilliseconds) {
+              ettMillisecondsFromNow = FMCMainDisplay.MILISECONDS_IN_DAY - (zuluTimeMilliseconds - ettInMilliseconds);
+            } else {
+              ettMillisecondsFromNow = ettInMilliseconds - zuluTimeMilliseconds;
+            }
+
+            // ETT cannot be more than 20 hours ahead of present time.
+            if (ettMillisecondsFromNow > 72000000) {
+              this.setScratchpadMessage(NXSystemMessages.entryOutOfRange);
+            } else {
+              this.flightPlanService.setPerformanceData(
+                'estimatedTakeoffTime',
+                Math.round(currentTimeStamp + ettMillisecondsFromNow),
+                forPlan,
+              );
+              this.flightPlanService.setPerformanceData('estimatedTakeoffTimeExpired', false, forPlan);
+              if (forPlan === FlightPlanIndex.Active && this.ettCheckSub.isPaused) {
+                this.ettCheckSub.resume();
+              }
+            }
+          }
+        } else {
+          this.setScratchpadMessage(NXSystemMessages.formatError);
+        }
+      }
+    }
+  }
+
+  /**
+   * Formats a time prediction based on seconds from present time, fms flightphase and ETT from the specified flightplan.
+   * @param secondsFromPresent seconds from present time
+   * @param forPlan plan to check ETT against.
+   * @returns formatted time prediction string in hhmm
+   */
+  getTimePrediction(secondsFromPresent: number, forPlan: FlightPlanIndex): string {
+    const phase = this.getFlightPhase();
+    const plan = this.flightPlanService.get(forPlan);
+    let time: number | null = null;
+    if (plan.isActiveOrCopiedFromActive() && phase > FmgcFlightPhase.Preflight && phase < FmgcFlightPhase.Done) {
+      time = this.simTime.get();
+    } else {
+      const ett = plan.performanceData.estimatedTakeoffTime.get();
+      if (ett !== null) {
+        // Use current time if ETT has expired
+        time = plan.performanceData.estimatedTakeoffTimeExpired.get() ? this.simTime.get() : ett;
+      }
+    }
+    if (time !== null) {
+      return FmsFormatters.milisecondsTohhmm(time + secondsFromPresent * 1000);
+    } else {
+      return FmsFormatters.secondsTohhmm(secondsFromPresent);
+    }
+  }
+
+  /**
+   * Gets the header to be used in pages for time predictions.
+   * @returns TIME if in flight or if ETT exists during preflight, otherwise UTC
+   */
+  getTimePredictionHeader(forPlan: FlightPlanIndex): 'UTC' | 'TIME' {
+    const phase = this.getFlightPhase();
+    const plan = this.flightPlanService.get(forPlan);
+    const ett = plan.performanceData.estimatedTakeoffTime.get();
+    if (
+      (plan.isActiveOrCopiedFromActive() && phase > FmgcFlightPhase.Preflight && phase < FmgcFlightPhase.Done) ||
+      ett !== null
+    ) {
+      //TODO check for valid clock signal
+      return 'UTC';
+    } else {
+      return 'TIME';
+    }
+  }
+
+  /**
+   * Checks if the ETT of the active flightplan has expired compared to zulu time.
+   * If it has, the associated flag is set on the performance data and the periodic interval check is paused.
+   * @param time, if specified, uses this time instead of the sim time
+   */
+  private checkEttExpired(time?: number): boolean {
+    const ett = this.flightPlanService.active.performanceData.estimatedTakeoffTime.get();
+    if (ett !== null && !this.flightPlanService.active.performanceData.estimatedTakeoffTimeExpired.get()) {
+      const isExpired = ett < (time ?? this.simTime.get());
+      if (isExpired) {
+        this.flightPlanService.setPerformanceData('estimatedTakeoffTimeExpired', isExpired, FlightPlanIndex.Active);
+        this.ettCheckSub.pause();
+      }
+      return isExpired;
+    }
+    return false;
+  }
+
+  /**
+   * Clears the ETT from the a flightplan
+   */
+  private clearEtt(forPlan: FlightPlanIndex = FlightPlanIndex.Active): void {
+    this.flightPlanService.setPerformanceData('estimatedTakeoffTime', null, forPlan);
+    this.flightPlanService.setPerformanceData('estimatedTakeoffTimeExpired', false, forPlan);
+    if (forPlan === FlightPlanIndex.Active) {
+      this.ettCheckSub.pause();
+    }
   }
 }