+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy/LegacyFmsPageInterface.ts
@@ -218,7 +218,10 @@ interface LegacyFmsPageFmsInterface extends FmsDataInterface, FmsDisplayInterfac
   setV2Speed(speed: number, forPlan: FlightPlanIndex): void;
   activateSecondaryPlan(index: number): Promise<void>;
   swapActiveAndSecondaryPlan(index: number): Promise<void>;
+  setEstimatedTakeoffTime(text: string, forPlan: FlightPlanIndex): void;
   computeAlternateCruiseLevel(forPlan: FlightPlanIndex): number | undefined;
+  getTimePrediction(secondsFromPresent: number, forPlan: FlightPlanIndex): string;
+  getTimePredictionHeader(forPlan: FlightPlanIndex): string;
 
   flightPlanService: FlightPlanService;
   navigationDatabase: NavigationDatabase;