+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/A320_Neo_CDU_FlightPlanPage.ts
@@ -1,5 +1,5 @@
 // @ts-strict-ignore
-// Copyright (c) 2021-2023, 2025 FlyByWire Simulations
+// Copyright (c) 2021-2023, 2025-2026 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
 import { A32NX_Util } from '../../../../shared/src/A32NX_Util';
@@ -11,8 +11,7 @@ import { CDUHoldAtPage } from './A320_Neo_CDU_HoldAtPage';
 import { AltitudeDescriptor, NXUnits, WaypointConstraintType } from '@flybywiresim/fbw-sdk';
 import { Keypad } from '../legacy/A320_Neo_CDU_Keypad';
 import { LegacyFmsPageInterface } from '../legacy/LegacyFmsPageInterface';
-import { FlightPlanLeg, isDiscontinuity } from '@fmgc/flightplanning/legs/FlightPlanLeg';
-import { FmsFormatters } from '../legacy/FmsFormatters';
+import { FlightPlanLeg, isDiscontinuity, isLeg } from '@fmgc/flightplanning/legs/FlightPlanLeg';
 import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
 import { Column, FormatLine } from '../legacy/A320_Neo_CDU_Format';
 
@@ -68,9 +67,6 @@ export class CDUFlightPlanPage {
         CDUFlightPlanPage.ShowPage(mcdu, offset, forPlan);
       }
     }, mcdu.PageTimeout.Medium);
-
-    const flightPhase = mcdu.flightPhaseManager.phase;
-    const isFlying = flightPhase >= FmgcFlightPhase.Takeoff && flightPhase != FmgcFlightPhase.Done;
 
     let showFrom = false;
 
@@ -336,21 +332,22 @@ export class CDUFlightPlanPage {
         }
 
         // Time
+        const flightplan = mcdu.flightPlanService.get(forPlan);
         let timeCell: string = Time.NoPrediction;
+        const isEttValid =
+          flightplan.performanceData.estimatedTakeoffTime.get() !== null &&
+          !flightplan.performanceData.estimatedTakeoffTimeExpired.get();
         let timeColor = 'white';
         if (verticalWaypoint && isFinite(verticalWaypoint.secondsFromPresent)) {
-          const utcTime = SimVar.GetGlobalVarValue('ZULU TIME', 'seconds');
+          timeCell = `${isFromLeg ? '{big}' : '{small}'}${mcdu.getTimePrediction(
+            verticalWaypoint.secondsFromPresent,
+            forPlan,
+          )}{end}`;
 
-          timeCell = `${isFromLeg ? '{big}' : '{small}'}${
-            isFlying
-              ? FmsFormatters.secondsToUTC(utcTime + verticalWaypoint.secondsFromPresent)
-              : FmsFormatters.secondsTohhmm(verticalWaypoint.secondsFromPresent)
-          }{end}`;
-
-          timeColor = color;
+          timeColor = isFromLeg && isEttValid ? 'magenta' : color;
         } else if (!inAlternate && fpIndex === targetPlan.originLegIndex) {
-          timeCell = '{big}0000{end}';
-          timeColor = color;
+          timeCell = `{big}${mcdu.getTimePrediction(0, forPlan)}{end}`;
+          timeColor = isEttValid ? 'magenta' : color;
         }
 
         // Fix Header
@@ -645,12 +642,9 @@ export class CDUFlightPlanPage {
         let timeCell: string = Time.NoPrediction;
         let timeColor = 'white';
         if (!shouldHidePredictions && Number.isFinite(pwp.flightPlanInfo.secondsFromPresent)) {
-          const utcTime = SimVar.GetGlobalVarValue('ZULU TIME', 'seconds');
           timeColor = color;
 
-          timeCell = isFlying
-            ? `${FmsFormatters.secondsToUTC(utcTime + pwp.flightPlanInfo.secondsFromPresent)}[s-text]`
-            : `${FmsFormatters.secondsTohhmm(pwp.flightPlanInfo.secondsFromPresent)}[s-text]`;
+          timeCell = `${mcdu.getTimePrediction(pwp.flightPlanInfo.secondsFromPresent, forPlan)}[s-text]`;
         }
 
         let speed: string = Speed.NoPrediction;
@@ -954,16 +948,16 @@ export class CDUFlightPlanPage {
 
           const timeRemaining = fmsGeometryProfile.getTimeToDestination();
           if (Number.isFinite(timeRemaining)) {
-            const utcTime = SimVar.GetGlobalVarValue('ZULU TIME', 'seconds');
-
-            destTimeCell = isFlying
-              ? FmsFormatters.secondsToUTC(utcTime + timeRemaining)
-              : FmsFormatters.secondsTohhmm(timeRemaining);
+            destTimeCell = mcdu.getTimePrediction(timeRemaining, forPlan);
           }
         }
       }
 
-      destText[0] = ['\xa0DEST', 'DIST\xa0\xa0EFOB', isFlying ? '\xa0UTC{sp}{sp}{sp}{sp}' : 'TIME{sp}{sp}{sp}{sp}'];
+      destText[0] = [
+        '\xa0DEST',
+        'DIST\xa0\xa0EFOB',
+        `${mcdu.getTimePredictionHeader(forPlan).padStart(4, '\xa0')}{sp}{sp}{sp}{sp}`,
+      ];
       destText[1] = [
         destCell,
         `{small}${destDistCell.padStart(4, '\xa0')}\xa0${destEFOBCell.padStart(5, '\xa0')}{end}`,
@@ -982,17 +976,19 @@ export class CDUFlightPlanPage {
         5,
         () => mcdu.getDelaySwitchPage(),
         () => {
-          CDUVerticalRevisionPage.ShowPage(
-            mcdu,
-            targetPlan.destinationLeg,
-            targetPlan.destinationLegIndex,
-            undefined,
-            undefined,
-            undefined,
-            undefined,
-            forPlan,
-            false,
-          );
+          if (isLeg(targetPlan.destinationLeg)) {
+            CDUVerticalRevisionPage.ShowPage(
+              mcdu,
+              targetPlan.destinationLeg,
+              targetPlan.destinationLegIndex,
+              undefined,
+              undefined,
+              undefined,
+              undefined,
+              forPlan,
+              false,
+            );
+          }
         },
       );
     }
@@ -1048,7 +1044,7 @@ export class CDUFlightPlanPage {
     }
     mcdu.setArrows(allowScroll, allowScroll, true, true);
     scrollText[0][1] = 'SPD/ALT\xa0\xa0\xa0';
-    scrollText[0][2] = isFlying ? '\xa0UTC{sp}{sp}{sp}{sp}' : 'TIME{sp}{sp}{sp}{sp}';
+    scrollText[0][2] = `${mcdu.getTimePredictionHeader(forPlan).padStart(4, '\xa0')}{sp}{sp}{sp}{sp}`;
 
     if (!showFrom) {
       fromHeader.update('');